fix: An instance alias appears in the database, Navigation componentization of log libraries
